Tue, Mar 19, 2024, 10:45 PM +08:00

+++ recover/recover.c
@@ -1,11 +1,12 @@
 #include <stdio.h>
 #include <stdlib.h>
+#include <stdint.h>
 
 int main(int argc, char *argv[])
 {
     //typedef *char[512] block;
     //typedef int64_t BYTE64; //64 * 8 = 512
-    int *buffer = malloc(8 * int64_t); // buffer is 512 bytes
+    int *buffer = malloc(8 * int8_t); // buffer is 512 bytes
 
     if (argc != 2)
     {
